Criando factory responsavel por instanciar a classe correta atraves do status cliente

+++ Principios-SOLID/CalculaDescontoFidelidade.cs
@@ -5,8 +5,8 @@
         public decimal AplicarDescontoFidelidade(decimal preco, int tempoDeContaEmAnos)
         {
             decimal porcentangemDescontoPorFidelidade = tempoDeContaEmAnos > Constantes.DESCONTO_MAXIMO_POR_FIDELIDADE ?
-                  Constantes.DESCONTO_MAXIMO_POR_FIDELIDADE / 100 :
-                  tempoDeContaEmAnos / 100;
+                  (decimal)Constantes.DESCONTO_MAXIMO_POR_FIDELIDADE / 100 :
+                  (decimal)tempoDeContaEmAnos / 100;
 
             return preco - (porcentangemDescontoPorFidelidade * preco);
         }
